pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-stock-briefing"
description = "Daily delta briefing for a fixed stock watchlist"
readme = "README.md"
requires-python = ">=3.12"
dynamic = ["version"]
dependencies = [
  "httpx>=0.27",
  "jinja2>=3.1",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
  "yfinance>=0.2.40",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "daily_stock_briefing.__version__"}
